完成前两个接口

+++ src/main/java/cn/edu/xidian/tafei_mall/model/vo/PromotionCreateVO.java
@@ -3,13 +3,14 @@ package cn.edu.xidian.tafei_mall.model.vo;
 import lombok.Data;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
 @Data
 public class PromotionCreateVO {
     private List<String> productIds;
-    private double discountRate;
+    private BigDecimal discountRate;
     private String startDate;
     private String endDate;
 }
